move coordinator and signer mods into subdirs
